Finish core navigation for runner prompt
